Fixed some jQuery for working category data
Fixed the fetching of the category name by removing the creation of
multiple click event handlers for the remove buttons after the categories.
Each execution of the updateEvents function would cause another click handler
to be made for each remove button. To solve this I removed the function and
assign an event handler to the button as the (button)/(category row) is created.

I also did a bit of code cleanup.