Grade calculator now calculates and is minified
I have added the ability to calculate the grade. This closes issue #004.
I also added a function to check to see that the categories add up to 100%.
The calculate function also makes sure that there is at least one grade for each
category to prevent div 0 errors.  Some CSS was added for the grade and
the position of the grade was changed as well. The javascript has been
minified and placed into gradecalc.min.js file. This closes issue #011.