Fixed input capture issue and empty grades issue
The first issue I fixed was when you selected a box that used
a regular expression to check it's contents on a blur event would not
let you deselect it. This was to to the fact that the field was NULL
and that was not an option as a permitted value. So it would capture
your DOM selection till you put something in there. To fix this I
added and not NULL logic to the blur event. This happend on the
category percentage and on the grade for an assignment.

The second issue occured when the user would click on the calculate
button and there were no grades entered in the input field but the
category would be selected. You would be given a grade of NaN
(Not a Number). To fix this I checked for NaN anytime a value is
retrieved. I also count the number of grades that are not NaN.
If the total is zero, I present the user a message to put some
grades in.